+++ CHANGELOG.md
@@ -10,6 +10,8 @@
     - **Features**: Add new methods to manage the STACKIT Git: `CreateInstance`, `DeleteInstance`, `GetInstance`
   - [v0.3.0](services/git/CHANGELOG.md#v030-2025-04-22)
     - **Features**: Add waiters to manage the STACKIT Git
+- `observability`: [v0.5.0](services/observability/CHANGELOG.md#v050-2025-04-16)
+  - **Feature:** Add new methods `ListLogsAlertgroups`, `CreateLogsAlertgroups`, `GetLogsAlertgroup`, `UpdateLogsAlertgroup`, `DeleteLogsAlertgroup`
 
 ## Release (2025-04-09)
 - `cdn`: [v0.3.0](services/cdn/CHANGELOG.md#v030-2025-04-04)